refactor: Add domain events.
1. Change tweet fetching error related notification base to TweetFetchingFailedNotificationEvent.

In tweet fetching failed notification context, only the tweet id was recorded
in notification id.
There is only one interaction(view tweet) for user and we can navigate user to the tweet with only tweet id.

2. Add UnknownNotificationButtonClickedEvent.

In normal situation, this event might never be used.
This event was added for debug.